readded sharing button
minor reformattings

share button position modified
